Add ConstituentModel as base for components and buses; removed SignalNode from evaluator tree
